Removed beerStyle property from BKSBeerDetailVC. Replaced with BeerStyle object's style.styleName
Enabled clicking of style button from DetailVC to go to Style Page. I decided to pass NSArray All Beers to the BKSBeerDetailVC so that NSArray beersOfStyle could be passed from BKSBeerDetailVC to BKSStyleVC, but I don't think this is necessarily the way we want to handle things. I'll wait for your input.